requirements.txt
Flask==2.0.1
flask-cors==3.0.10
flask-limiter==2.8.1
Pillow==8.4.0
--find-links https://download.pytorch.org/whl/cpu/torch_stable.html
torch==1.9.0+cpu
torchvision==0.10.0+cpu